proposed refactoring
